refactor Write, Update 중복 내용 리팩토링 * Category API 관련, useCategoryAPI 생성 및 이관
