+++ frontend/src/App.js
@@ -9,7 +9,7 @@ import Login from './Components/Login/Login'
 import BoardDetail from './Components/BoardDetail/BoardDetail'
 import Write from './Components/Write/Write'
 import Profile from './Components/Profile/Profile'
-import Update from './Components/Update/Update'
+import Update from './Components/Write/Update'
 import Footer from './Components/Footer/Footer'
 import { UserProvider } from 'userContext'
 import ProfileUpdate from './Components/ProfileUpdate/ProfileUpdate'
